Move fetchValidVersions to server.ts file; update imports
